project/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
email_validator==2.2.0
marshmallow==3.25.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyMySQL==1.1.1
Flask-JWT-Extended==4.7.1
Flask-Caching==2.3.0
redis==5.2.1
marshmallow-sqlalchemy==1.4.1
scipy==1.15.2
itsdangerous>=2.0.0
